+++ Tests/Integration/NonTransacting/Network/Connection/FogBlockConnectionIntTests.swift
@@ -213,10 +213,8 @@ extension FogBlockConnectionIntTests {
     func createFogBlockConnection(networkConfig: NetworkConfig) -> FogBlockConnection {
         let httpFactory = HttpProtocolConnectionFactory(
                 httpRequester: networkConfig.httpRequester ?? DefaultHttpRequester())
-        let grpcFactory = GrpcProtocolConnectionFactory()
         return FogBlockConnection(
             httpFactory: httpFactory,
-            grpcFactory: grpcFactory,
             config: networkConfig,
             targetQueue: DispatchQueue.main)
     }