+++ Tests/Integration/NonTransacting/Network/Connection/FogReportConnectionIntTests.swift
@@ -71,10 +71,8 @@ extension FogReportConnectionIntTests {
     ) throws -> FogReportConnection {
         let url = try FogUrl.make(string: IntegrationTestFixtures.network.fogReportShortUrl).get()
         let httpFactory = HttpProtocolConnectionFactory(httpRequester: DefaultHttpRequester())
-        let grpcFactory = GrpcProtocolConnectionFactory()
         return FogReportConnection(
             httpFactory: httpFactory,
-            grpcFactory: grpcFactory,
             url: url,
             transportProtocolOption: transportProtocol.option,
             targetQueue: DispatchQueue.main)
@@ -85,10 +83,8 @@ extension FogReportConnectionIntTests {
     ) throws -> FogReportConnection {
         let url = try FogUrl.make(string: IntegrationTestFixtures.network.fogReportUrl).get()
         let httpFactory = HttpProtocolConnectionFactory(httpRequester: DefaultHttpRequester())
-        let grpcFactory = GrpcProtocolConnectionFactory()
         return FogReportConnection(
             httpFactory: httpFactory,
-            grpcFactory: grpcFactory,
             url: url,
             transportProtocolOption: transportProtocol.option,
             targetQueue: DispatchQueue.main)